+++ bin/cli/Cargo.toml
@@ -22,6 +22,6 @@ clap                           = { workspace = true }
 futures                        = { workspace = true }
 hex                            = { workspace = true }
 miden-objects                  = { workspace = true }
-miden-private-transport-client = { features = ["sqlite", "tonic"], workspace = true }
+miden-private-transport-client = { features = ["sqlite", "testing", "tonic"], workspace = true }
 tokio                          = { workspace = true }
 tracing                        = { workspace = true }
