+++ crates/node/src/lib.rs
@@ -1,4 +1,4 @@
-//! # Miden Transport Layer Node Library
+//! # Miden Transport Layer Client Library
 //!
 //! Implementation of the Miden Transport Layer node for private notes.
 //!
@@ -12,8 +12,8 @@
 //! - metrics and traces, exported through the OpenTelemetry framework for monitoring.
 //!
 //! ## Database
-//! Notes are stored in a database, implementing the [`Database`](`crate::database::DatabaseBackend`).
-//! A SQLite-based  implementation is provided.
+//! Notes are stored in a database, implementing the
+//! [`Database`](`crate::database::DatabaseBackend`). A SQLite-based  implementation is provided.
 //!
 //! ### Maintenance
 //! A periodic task [`DatabaseMaintenance`](`crate::database::DatabaseMaintenance`) takes care of
@@ -21,15 +21,15 @@
 //!
 //! ## Telemetry
 //! Metrics and traces to monitor the node state are provided.
-//! While metrics provide insights into general requests stats, traces can provide insights into specific
-//! requests.
+//! While metrics provide insights into general requests stats, traces can provide insights into
+//! specific requests.
 //! Metrics and traces can be exported following using the [OpenTelemetry](https://opentelemetry.io) framework.
 
 #![deny(missing_docs)]
 
 /// Database
 pub mod database;
-/// Main error type employed
+/// Error management
 pub mod error;
 /// Tracing, metrics export configuration
 pub mod logging;
@@ -37,13 +37,13 @@ pub mod logging;
 pub mod metrics;
 /// Main node implementation
 pub mod node;
-/// Types used
-pub mod types;
 /// Testing functions
 ///
 /// Gated through the `testing` feature.
 #[cfg(feature = "testing")]
 pub mod test_utils;
+/// Types used
+pub mod types;
 
 pub use error::{Error, Result};
 pub use node::{Node, NodeConfig, grpc::GrpcServer};
